fix(cli): get git path for whole project instead for each file (#159)
* fix(cli): get git path for whole project instead for each file

Before this fix the git path was received through the `git` command for
every file. Each `gitPath` resolution needed ~25-30ms on my machine.
This PR changes the `gitPath` resolution to be only received at startup
by the `git` command. Then this git project project path is used to
resolve the git path for each file.

In addition this PR solves a bug that happenend on the material-ui
repository:
Their webpack output for files were missing the `./` prefix. This caused
the dependency arrays in the stories file to be empty.

* Fix tests and resolve accidentally pushed merge conflicts

Co-authored-by: Johannes Merz <JohannesMerz@users.noreply.github.com>